Fix Template initialization and comparison.
Previously we didn't add the AST to the Template if errors were
present. This behavior probably wasn't intuitive, so this commit
improves initialization to always add the AST nodes.

Also, fix a bug where comparison of Template succeeded even if error
messages were different. This was because Templates were only compared
based on AST, and not contents of the errors or remaining tokens.

Closes #7.