Remove 'send' in implementation in favor of substitutions from a Hash
Previously, we used `public_send` to figure out what text should be
substituted in a template after it had been parsed. This was
concerning to me because:

* Dispatching arbitrary methods via `send` or `public_send` should only be done sparingly, due to the possibility of opening up RCE vulnerabilities.
* The previous flow led to highly-coupled code with bidirectional dependencies - the caller of the templating system was then called via methods defined by the whitelisted placeholders. So previously, the templating system was a dependency of the calling code, and the calling code was a dependency of the templating system!

By changing the implementation to just pass a Hash containing keys
with placeholder names, and values containing the substitutions, we
make the library easier to use safely, and less coupled.

One side benefit (which may not be worth pursuing) is that we could
potentially replace this library in the future with a native
implementation, or even a microservice. However if we keep the current
implementation, it seems to me that it would be hard to have a native
library call back to the caller, or in the micro-service case it would
require setting up an API on the caller to receive queries from the
templating engine. I think that the complexity of extracting this out
shows weaknesses in the previous implementation which are addressed by
this simple change.